test_neuronoptions: run as standalone script
